Petal/requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
networkx
pandas
matplotlib
langgraph
langsmith
whisper
SpeechRecognition
PyAudio
deep_translator
tiktoken
duckduckgo-search
